+++ WatchCloud Watch App/Extension/Color.swift
@@ -12,4 +12,8 @@ extension Color {
     static var cellBG: Color {
         .gray.opacity(0.2)
     }
+    
+    static var offBlack: Color {
+        Color(uiColor: UIColor(white: 0.1, alpha: 1))
+    }
 }