+++ WatchCloud Watch App/UI/SettingsView.swift
@@ -23,6 +23,7 @@ struct SettingsView: View {
         List {
             playbackSection
             downloadsSection
+            appInfoSection
         }
         .alert("Are you sure you want to remove all downloads?", isPresented: $showDeleteAllAlert) {
             Button("Remove All", role: .destructive) { try? audioStore.removeAllDownloads() }
@@ -50,7 +51,7 @@ struct SettingsView: View {
     private var downloadsSection: some View {
         if Config.isDownloadingEnabled(for: userStore.myUser?.id) {
             Section(
-                header: Text("Downloads")
+                header: Text("Downloads").padding(.bottom, 2)
             ) {
                 Toggle(isOn: Config.$allowDownloadingUsingData, label: {
                     Text(String(localized: "Cellular Data", comment: "Toggle label"))
@@ -79,6 +80,24 @@ struct SettingsView: View {
                     .listItemTint(.clear)
                 }
             }
+        }
+    }
+    
+    private var appInfoSection: some View {
+        Section(
+            header: Text("App").padding(.bottom, 2)
+        ) {
+            HStack {
+                Text("Version")
+                Spacer()
+                Text(Config.appVersion)
+                    .fontWeight(.medium)
+            }
+            
+            PoweredBySCView()
+                .fullWidth()
+                .padding(.top, 6)
+                .listRowBackground(Color.clear)
         }
     }
 }